Sync resources.json with that in gem5/gem5-resources
For reasons that are not presently understood, the resources.json file
hosted at http://resources.gem5.org/resource.json is out-of-sync with
that found in the root of the https://github.com/gem5/gem5-resources
repository. The resources.json file hosted on the website appears to be
missing entires.

This is known to produce errors, such as that in gem5/gem5#127.

This patch syncs the resources in gem5/gem5-resources to that hosted on
the website. This should fix these errors.